lifecap.sh: added -a flag (archive)

+++ lifecap.sh
@@ -13,6 +13,30 @@ vcodec=libx264
 lockf=/tmp/lifecap.lock
 pid=$$
 
+archive() {
+    cd $outputdir || exit 1
+    now=(`date +'%Y %m'`)
+    for i in *.mp4; do
+        v=(`echo $i | sed -r 's/(....)(..).*/\1 \2/'`)
+        if [ ${v[0]} = ${now[0]} ] && [ ${v[1]} = ${now[1]} ]; then
+            continue
+        fi
+        dir="${v[0]}/${v[1]}"
+        mkdir -pv $dir || exit 1
+        mv -v $i $dir
+    done
+}
+
+if [ "$1" = '-a' ]; then
+    archive
+    exit 0
+fi
+
+if [ $# != 0 ]; then
+    printf "usage: %s [-a]\n" $0 >&2
+    exit 64
+fi
+
 trap "rm -f $lockf" EXIT
 
 (
